requirements.txt
requests
beautifulsoup4
markdownify
datetime
gitpython
